ai_job_agent/requirements.txt
google-search-results
python-dotenv
tenacity
pytest
pyyaml
spacy
google-genai
pdfplumber
